Fix BUG in a class diagram PlantUML to BUML transformation.
The BUG occurs when an association have the same source and target class. It raises the error: ValueError: Association ends must have unique names

To fix it, I simply check if both ends of the association have the same Class name and rename them to Class_sorce and Class_target.